add tests of functions meant to be kept
